+++ public/tier1/strtools.h
@@ -22,6 +22,9 @@
 #include <string.h>
 #include <stdlib.h>
 
+#include "utlleanvector.h"
+#include "utlvector.h"
+
 // Forward declaration
 class CBufferString;
 class Vector;
@@ -32,9 +35,6 @@ class Quaternion;
 class Color;
 class QAngle;
 class CUtlString;
-
-template< class T, class I> class CUtlVectorMemory;
-template< class T, class I, class A> class CUtlVector;
 
 abstract_class IParsingErrorListener
 {
@@ -692,17 +692,6 @@ PLATFORM_INTERFACE bool _V_StrSubst( const char *pIn, const char *pMatch, const 
 									 OUT_Z_CAP( outLen ) char *pOut, int outLen, bool bCaseSensitive=false );
 #define V_StrSubst _V_StrSubst
 
-// AM TODO: If possible, use CSplitString instead rn. 
-// These are exported by tier0, but will require changes to CUtlVector (additional template arg)
-// 
-// Split the specified string on the specified separator.
-// Returns a list of strings separated by pSeparator.
-// You are responsible for freeing the contents of outStrings (call outStrings.PurgeAndDeleteElements).
-PLATFORM_OVERLOAD void V_SplitString( const char *pString, const char *pSeparator, CUtlVector<CUtlString, int, CUtlVectorMemory<CUtlString, int>> &outStrings, bool include_empty = false );
-
-// Just like V_SplitString, but it can use multiple possible separators.
-PLATFORM_OVERLOAD void V_SplitStringInPlace( char *pString, const char *pSeparator, CUtlVector<const char *, int, CUtlVectorMemory<const char *, int>> &outStrings );
-
 // This function takes a slice out of pStr and stores it in pOut.
 // It follows the Python slice convention:
 // Negative numbers wrap around the string (-1 references the last character).
@@ -1074,5 +1063,92 @@ PLATFORM_INTERFACE void V_CopyMemory3D(
 #define Q_FixupPathName			V_FixupPathName
 
 #endif // !defined( VSTDLIB_DLL_EXPORT )
+
+// Split the specified string on the specified separator.
+// Returns a list of strings separated by pSeparator.
+// You are responsible for freeing the contents of outStrings (call outStrings.PurgeAndDeleteElements).
+PLATFORM_OVERLOAD void V_SplitString( const char *pString, const char *pSeparator, CUtlVector<CUtlString> &outStrings, bool include_empty = false );
+
+// Just like V_SplitString, but it can use multiple possible separators.
+PLATFORM_OVERLOAD void V_SplitStringInPlace( char *pString, const char *pSeparator, CUtlLeanVector<const char *> &outStrings );
+
+// easy string list class with dynamically allocated strings. For use with V_SplitString, etc.
+// Frees the dynamic strings in destructor.
+class CUtlStringList : public CUtlVectorAutoPurge< char * >
+{
+public:
+	~CUtlStringList()
+	{
+		PurgeAndDeleteElements();
+	}
+
+	void CopyAndAddToTail( char const *pString )			// clone the string and add to the end
+	{
+		char *pNewStr = new char[1 + strlen( pString )];
+		V_strcpy( pNewStr, pString );
+		AddToTail( pNewStr );
+	}
+
+	static int __cdecl SortFunc( char *const *sz1, char *const *sz2 )
+	{
+		return strcmp( *sz1, *sz2 );
+	}
+
+	inline void PurgeAndDeleteElements()
+	{
+		for(int i = 0; i < m_Size; i++)
+		{
+			delete[] Element( i );
+		}
+		Purge();
+	}
+};
+
+class CSplitString : public CUtlVector< char * >
+{
+public:
+	// Splits the string based on separator provided
+	// Example: string "test;string string2;here" with a separator ";"
+	// results in [ "test", "string string2", "here" ] array entries.
+	// Separator can be multicharacter, i.e. "-;" would split string like
+	// "test-;string" into [ "test", "string" ]
+	CSplitString( const char *pString, const char *pSeparator, bool include_empty = false ) : m_szBuffer( nullptr ), m_nBufferSize( 0 )
+	{
+		Split( pString, -1, &pSeparator, 1, include_empty );
+	}
+
+	// Splits the string based on array of separators provided
+	// Example: string "test;string,string2;here" with a separators [ ";", "," ]
+	// results in [ "test", "string", "string2", "here" ] array entries.
+	// Separator can be multicharacter, i.e. "-;" would split string like
+	// "test-;string" into [ "test", "string" ]
+	CSplitString( const char *pString, const char **pSeparators, int nSeparators, bool include_empty = false ) : m_szBuffer( nullptr ), m_nBufferSize( 0 )
+	{
+		Split( pString, -1, pSeparators, nSeparators, include_empty );
+	}
+
+	~CSplitString()
+	{
+		if(m_szBuffer)
+			delete[] m_szBuffer;
+	}
+
+	// Works the same way as CSplitString::Split with the only exception that it allows you 
+	// to provide single string of separators like ";,:" which will then get used to separate string
+	// instead of providing an array of separators, but doesn't support multicharacter separators cuz of that!
+	DLL_CLASS_IMPORT void SetPacked( const char *pString, const char *pSeparators, bool include_empty = false, int stringSize = -1 );
+
+private:
+	DLL_CLASS_IMPORT void Split( const char *pString, int stringSize, const char **pSeparators, int nSeparators, bool include_empty = false );
+
+	void PurgeAndDeleteElements()
+	{
+		Purge();
+	}
+
+private:
+	char *m_szBuffer; // a copy of original string, with '\0' instead of separators
+	int m_nBufferSize;
+};
 
 #endif	// TIER1_STRTOOLS_H
